backend/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.27,<1
websockets>=12,<14
databricks-zerobus-ingest-sdk>=0.1
python-dotenv>=1,<2
httpx>=0.27,<1
